# services/auth-service/src/main/openapi/swagger.yaml
...
openapi: 3.0.1
info:
  title: Authentication and user API
  version: 1.0.0
  description: Authentication, user profiles, seller onboarding, role administration,
    and public signing-key discovery.
  contact:
    name: Ecommerce Platform maintainers
servers:
- url: http://localhost:8081
  description: Local auth-service direct access
paths:
...
  "/api/admin/users/{userId}/roles":
    put:
      tags:
      - Administration
      operationId: setRoles
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AssignRolesRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseUserProfileDto"
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Assign user roles
      description: Replaces the roles assigned to the specified user.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
  "/api/admin/users/{userId}/active":
    put:
      tags:
      - Administration
      operationId: setActive
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: active
        in: query
        required: true
        schema:
          type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseUserProfileDto"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Set user active status
      description: Activates or deactivates the specified user account.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
  "/api/admin/seller-applications/{id}/approve":
    put:
      tags:
      - Administration
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseSellerApplicationResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Approve seller application
      description: Approves the specified seller application.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
  "/api/admin/roles/{roleId}/permissions":
    put:
      tags:
      - Administration
      operationId: updateRolePermissions
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateRolePermissionsRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseRole"
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Update role permissions
      description: Replaces the permissions assigned to the specified role.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
  "/api/admin/roles":
    get:
      tags:
      - Administration
      operationId: listRoles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseListRole"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: List roles
      description: Returns configured roles.
      security:
      - bearerAuth: []
    post:
      tags:
      - Administration
      operationId: createRole
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateRoleRequest"
        required: true
      responses:
        '201':
          description: Role created.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseRole"
        '400':
          description: Invalid role name.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Role already exists.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Create a role
      description: Creates a role with the supplied name.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
  "/api/admin/users":
    get:
      tags:
      - Administration
      operationId: listUsers
      x-spring-paginated: true
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
        description: Zero-based page index.
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          minimum: 1
        description: Maximum number of results per page.
      - name: sort
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
        description: Sort criteria in property,direction form; repeat to add sort
          keys.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponsePageUserProfileDto"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: List users
      description: Returns a paginated list of users.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
  "/api/admin/users/{userId}":
    get:
      tags:
      - Administration
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ApiResponseUserProfileDto"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Get user by ID
      description: Returns the specified user profile.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
  "/api/admin/roles/{roleId}":
    delete:
      tags:
      - Administration
      operationId: deleteRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Role deleted.
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: "Bearer-token filter rejection: revoked, non-user, or invalid-identity tokens return 401; the body may be empty, and revoked tokens may return plain text. MVC-level authentication failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
            text/plain:
              schema:
                type: string
        '403':
          description: "Security denial: anonymous requests receive 403 with an empty body when OAuth2 login is disabled, and callers without permission receive 403. With OAuth2 login enabled and a provider configured, anonymous access may instead redirect to sign-in (302). MVC-level authorization failures use a JSON ErrorResponse."
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '302':
          description: Conditional sign-in redirect when OAuth2 login is enabled and a provider is configured; otherwise anonymous protected requests receive 403.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '503':
          description: "When a Bearer token is present and the revocation blacklist check fails, JwtAuthFilter rejects the request with an empty response."
      summary: Delete a role
      description: Deletes the specified role.
      security:
      - bearerAuth: []
    servers:
    - url: http://localhost:8081
      description: Local auth-service direct access
    - url: http://localhost:8080
      description: Local API gateway
...
components:
  schemas:
...
    ApiResponseUserProfileDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/UserProfileDto"
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
...
    AssignRolesRequest:
      required:
      - roles
      type: object
      properties:
        roles:
          type: array
          x-not-null-message: must not be empty
          x-not-blank-items: true
          items:
            type: string
          minItems: 1
          x-size-message: must not be empty
...
    ApiResponseSellerApplicationResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/SellerApplicationResponse"
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
...
    UpdateRolePermissionsRequest:
      required:
      - permissions
      type: object
      properties:
        permissions:
          type: array
          x-not-null-message: must not be empty
          x-not-blank-items: true
          items:
            type: string
          minItems: 1
          x-size-message: must not be empty
    ApiResponseRole:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/Role"
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        permissions:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/Permission"
...
    CreateRoleRequest:
      required:
      - name
      type: object
      properties:
        name:
          pattern: "^ROLE_[A-Z_]+$"
          type: string
          x-not-null-message: must not be blank
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "must not be blank")'
          x-pattern-message: Role name must match ROLE_[A-Z_]+
    ApiResponsePageUserProfileDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/PageUserProfileDto"
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
    PageUserProfileDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/UserProfileDto"
        number:
          type: integer
          format: int32
        sort:
          "$ref": "#/components/schemas/SortObject"
        pageable:
          "$ref": "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    ApiResponseListRole:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/Role"
          nullable: true
          x-jackson-json-include-policy: ALWAYS
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
      - status
      - message
      - data
      - timestamp
    ErrorResponse:
      type: object
      description: Standard error body returned by the shared exception handler.
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        code:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...